examples/quickstart_tensorflow/pyproject.toml
[build-system]
requires = [
    "poetry==1.1.13",
]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "quickstart_tensorflow"
version = "0.1.0"
description = "Keras Federated Learning Quickstart with Flower"
authors = ["The Flower Authors <enquiries@flower.dev>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
# flwr = "^0.19.0"
# flwr = { path = "../../", develop = true }  # Development
flwr = { git = "https://github.com/achekerylla/flower.git", branch = "achekerylla-test" }
tensorflow-cpu = "^2.9.0"
